Refactor removeAccountDataHook to ease the testing
