+++ src/utils/hooks/setAccountDataHooks.ts
@@ -1,12 +1,21 @@
 import { useDispatch } from "react-redux";
 
 import {
+  useGetAccountsByFingerPrint,
+  useGetAccountsByType,
   useGetNextAvailableAccountLabels,
   useImplicitAccounts,
   useSeedPhrases,
 } from "./getAccountDataHooks";
 import { useSelectedNetwork } from "./networkHooks";
-import { LedgerAccount, MnemonicAccount, SocialAccount } from "../../types/Account";
+import { useRemoveAccountsDependencies } from "./removeAccountDependenciesHooks";
+import {
+  AccountType,
+  LedgerAccount,
+  MnemonicAccount,
+  SecretKeyAccount,
+  SocialAccount,
+} from "../../types/Account";
 import { makeDerivationPath } from "../account/derivationPathUtils";
 import { makeMnemonicAccount } from "../account/makeMnemonicAccount";
 import { decrypt, encrypt } from "../crypto/AES";
@@ -15,6 +24,8 @@ import { useAppDispatch } from "../redux/hooks";
 import { accountsSlice } from "../redux/slices/accountsSlice";
 import { restore as restoreFromSecretKey } from "../redux/thunks/secretKeyAccount";
 import { derivePublicKeyPair, getFingerPrint } from "../tezos";
+
+const { removeMnemonicAndAccounts, removeNonMnemonicAccounts } = accountsSlice.actions;
 
 const { addAccount } = accountsSlice.actions;
 
@@ -171,5 +182,57 @@ export const useRestoreSocial = () => {
       label,
     };
     dispatch(addAccount(account));
+  };
+};
+
+/**
+ * Hook for removing all accounts from mnemonic group by a given fingerprint.
+ */
+export const useRemoveMnemonic = () => {
+  const dispatch = useAppDispatch();
+  const getAccountsByFingerPrint = useGetAccountsByFingerPrint();
+  const removeAccountsDependencies = useRemoveAccountsDependencies();
+
+  return (fingerPrint: string) => {
+    removeAccountsDependencies(getAccountsByFingerPrint(fingerPrint));
+
+    dispatch(
+      removeMnemonicAndAccounts({
+        fingerPrint,
+      })
+    );
+  };
+};
+
+/**
+ * Hook for removing all accounts of a given type.
+ */
+export const useRemoveNonMnemonic = () => {
+  const dispatch = useAppDispatch();
+  const getAccountsByType = useGetAccountsByType();
+  const removeAccountsDependencies = useRemoveAccountsDependencies();
+
+  return (accountType: AccountType) => {
+    removeAccountsDependencies(getAccountsByType(accountType));
+
+    dispatch(
+      removeNonMnemonicAccounts({
+        accountType,
+      })
+    );
+  };
+};
+
+/**
+ * Hook for removing single account.
+ */
+export const useRemoveAccount = () => {
+  const dispatch = useAppDispatch();
+  const removeAccountsDependencies = useRemoveAccountsDependencies();
+
+  return (account: SocialAccount | LedgerAccount | SecretKeyAccount) => {
+    removeAccountsDependencies([account]);
+
+    dispatch(accountsSlice.actions.removeAccount(account));
   };
 };
