+++ src/utils/redux/slices/accountsSlice.ts
@@ -44,7 +44,7 @@ export const accountsSlice = createSlice({
   },
   reducers: {
     reset: () => initialState,
-    // Do not call this directly, use useRemoveMnemonic  from removeAccountDataHooks
+    // Do not call this directly, use useRemoveMnemonic  from setAccountDataHooks
     removeMnemonicAndAccounts: (
       state,
       { payload }: { type: string; payload: { fingerPrint: string } }
@@ -56,7 +56,7 @@ export const accountsSlice = createSlice({
       state.items = newAccounts;
       delete state.seedPhrases[fingerPrint];
     },
-    // Do not call this directly, use useRemoveNonMnemonic from removeAccountDataHooks
+    // Do not call this directly, use useRemoveNonMnemonic from setAccountDataHooks
     removeNonMnemonicAccounts: (
       state,
       { payload }: { type: string; payload: { accountType: AccountType } }
@@ -69,7 +69,7 @@ export const accountsSlice = createSlice({
         state.secretKeys = {};
       }
     },
-    // Do not call this directly, use useRemoveAccount from removeAccountDataHooks
+    // Do not call this directly, use useRemoveAccount from setAccountDataHooks
     removeAccount: (
       state,
       { payload }: { type: string; payload: SocialAccount | LedgerAccount | SecretKeyAccount }
